Restructured project directory
Move EmailTemplate to components\emails dir
NewUserForm is now its own component in components\forms
