+++ app/api/send/route.ts
@@ -1,4 +1,4 @@
-import { EmailTemplate } from '@/components/ui/email-template';
+import { EmailTemplate } from '@/components/emails/email-template';
 import { Resend } from 'resend';
 
 const resend = new Resend(process.env.RESEND_API_KEY);
