+++ components/emails/email-template.tsx
